BookCard is refactored. User follow bug fixed. UserDetailPage is removed
