+++ FlutterBookApp/lib/cubit/UserFollowCubit.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_book_app/httprequest/HttpRequestUser.dart';
+import 'package:flutter_book_app/httprequest/Model/ResponseEntity.dart';
 import 'package:flutter_book_app/model/dto/UserFriendDTO.dart';
 import 'package:logger/logger.dart';
 
@@ -15,12 +16,21 @@ class UserFollowProcessCubit extends Cubit<UserFollowProcessCubitData?> {
       UserFriendDTO userFriendDTO) async {
     data = UserFollowProcessCubitData(userFriendDTO: userFriendDTO);
     data.userIsFollowed = await HttpRequestUser.followUser(userFriendDTO.id);
-    print("userIsFollowed : ${data.userIsFollowed}");
     emit(data);
     return data;
   }
 
-  void unfollowUser(UserFriendDTO userFriendDTO) async {
+  Future<UserFollowProcessCubitData> unfollowUser(
+      UserFriendDTO userFriendDTO) async {
+    data = UserFollowProcessCubitData(userFriendDTO: userFriendDTO);
+    bool success = await HttpRequestUser.unfollowUser(userFriendDTO.id);
+    if(success){
+      data.userIsFollowed=false;
+    }else{
+      data.userIsFollowed=true;
+    }
+    emit(data);
+    return data;
     // bool result = await HttpRequestUser.followUser(userFriendDTO.id);
     // return result;
   }
