+++ FlutterBookApp/lib/product/BookCardOld.dart
@@ -1,63 +1,55 @@
+/*
+
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_book_app/model/Recommend.dart';
+import 'package:flutter_book_app/util/ResponsiveDesign.dart';
 import 'package:logger/logger.dart';
 import '../cubit/recommendedbook/BookCubit.dart';
-import '../httprequest/HttpRequestBook.dart';
 import '../model/Book.dart';
 import '../util/ProductColor.dart';
-import '../util/ResponsiveDesign.dart';
 import 'BookDesignDecoration.dart';
 
-class RecommendBookCard extends StatefulWidget {
-  late RecommendData recData;
+class BookCard extends StatefulWidget {
+  late Book book;
   late int index;
+  late bool isBookRead;
 
-  RecommendBookCard({super.key, required this.index, required this.recData});
+  BookCard({super.key, required this.book, required this.index});
 
   @override
-  State<RecommendBookCard> createState() => _RecommendBookCardState();
+  State<BookCard> createState() => _BookCardState();
 }
 
-class _RecommendBookCardState extends State<RecommendBookCard> {
+class _BookCardState extends State<BookCard> {
   var log = Logger(printer: PrettyPrinter(colors: false));
 
   final double imgWidth = ResponsiveDesign.width() / 5.5;
   final double imgHeight = ResponsiveDesign.height() / 6;
   final double padding = ResponsiveDesign.height() / 65;
-
   bool isLoading = true;
-  // List<Book> bookList = [];
 
-  @override
-  void initState() {
-    super.initState();
-    print("Recommended BookCard initState");
-    // _retrieveReadBookData();
+*/
+/*
+  _retrieveUserReadThisBook() async {
+    await _retrieveReadBookList();
+    setState(() {
+      isLoading = false;
+    });
   }
 
-  /*_retrieveReadBookData() async {
-    await _retrieveReadBookList();
-    if (mounted) {
-      setState(() {
-        isLoading = false;
-      });
-    } else {
-      print("Elseye girdi mounted: $mounted");
+  _retrieveReadBookList() async {
+    Book book = await HttpRequestBook.getIfUserReadBook(widget.book.id);
+    if (book.id == widget.book.id) {
+      widget.isBookRead = true;
     }
-    // setState(() {
-    //   isLoading = false;
-    // });
-  }*/
+  }*//*
 
-  /*_retrieveReadBookList() async {
-    bookList = await HttpRequestBook.getReadBookList();
-  }*/
 
   @override
   Widget build(BuildContext context) {
+    // _retrieveUserReadThisBook();
     return SizedBox(
-      height: imgHeight * 1.9,
+      height: imgHeight * 1.85,
       child: Column(
         children: [
           Row(
@@ -66,13 +58,12 @@ class _RecommendBookCardState extends State<RecommendBookCard> {
                 children: [
                   InkWell(
                     onTap: () {
-                      goToDetailPage(context, widget.recData.data);
+                      goToDetailPageOfBook(context, widget.book);
                     },
-                    child: getRecommendCardContent(),
+                    child: getBookCardContent(),
                   ),
                   Padding(
-                    padding:
-                        EdgeInsets.only(top: ResponsiveDesign.height() / 50),
+                    padding: EdgeInsets.only(top: ResponsiveDesign.height() / 50),
                     child: Row(
                       children: [
                         getBookImage(context),
@@ -89,83 +80,65 @@ class _RecommendBookCardState extends State<RecommendBookCard> {
     );
   }
 
-  /*bool isBookReadByUser(Book book) {
-    bool isBookRead = false;
-    for (var element in bookList) {
-      if (element.id == book.id) {
-        isBookRead = true;
-        continue;
-      }
-    }
-    log.i("Book is Read : $isBookRead");
-    return isBookRead;
-  }*/
-
-  Padding getRecommendCardContent() {
+  Padding getBookCardContent() {
     final double contentWidth = imgWidth / 2 + ResponsiveDesign.width() / 25;
     return Padding(
-      padding: EdgeInsets.only(left: contentWidth),
+      padding: EdgeInsets.only(
+        left: contentWidth,
+      ),
       child: ContainerWithBoxDecoration(
         child: Container(
           // width: 295,
-          /*width: ResponsiveDesign.width() - contentWidth - 4.5 * padding,
-          height: imgHeight + 5.5 * padding,*/
           width: ResponsiveDesign.width() - contentWidth - 5.5 * padding,
           height: imgHeight + 7 * padding,
           color: ProductColor.white,
           child: Padding(
-            padding: EdgeInsets.only(left: imgWidth, top: 5),
+            padding: EdgeInsets.only(left: imgWidth, top: 10),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                getShortTitle("${widget.index}-) ${widget.recData.data.name}"),
-                /*Text(
-                  getShortTitle(
-                      "${widget.index}-) ${widget.recData.data.name}"),
+                getShortTitle("${widget.index}-) ${widget.book.name}")
+                */
+/*Text(
+                  getShortTitle("${widget.index}-) ${widget.book.name}"),
                   maxLines: 2,
                   style: const TextStyle(
                       fontSize: 18, fontWeight: FontWeight.bold),
-                ),*/
+                )*//*
+
+                ,
                 Expanded(
                   child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      Padding(
-                        padding: const EdgeInsets.only(top: 5),
-                        child: Row(
-                          children: [
-                            getBookRatingShape(widget.recData.data.point),
-                          ],
-                        ),
+                    Padding(
+                      padding: const EdgeInsets.only(top: 10),
+                      child: Row(
+                        children: [
+                          getBookRatingShape(widget.book.point),
+                        ],
                       ),
-                      const SizedBox(
-                        height: 12,
+                    ),
+                    const SizedBox(
+                      height: 10,
+                    ),
+                    Text(
+                      "${widget.book.totalRead} Reviews",
+                      style: const TextStyle(
+                          fontSize:16,
+                          fontWeight: FontWeight.bold,
+                          color: ProductColor.grey),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.only(top: 10),
+                      child: Text(
+                        getShortDesc(widget.book.desc),
+                        maxLines: 2,
+                        style:
+                        const TextStyle(fontSize: 16, color: ProductColor.grey),
                       ),
-                      Text(
-                        "${widget.recData.data.totalRead} Reviews",
-                        style: const TextStyle(
-                            fontSize: 15,
-                            fontWeight: FontWeight.bold,
-                            color: ProductColor.grey),
-                      ),
-                      Text(
-                        widget.recData.by,
-                        style: TextStyle(
-                            fontSize: 15,
-                            fontWeight: FontWeight.bold,
-                            color: widget.recData.color),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.only(top: 5),
-                        child: Text(
-                          getShortDesc(widget.recData.data.desc),
-                          maxLines: 3,
-                          style: const TextStyle(
-                              fontSize: 15, color: ProductColor.grey),
-                        ),
-                      ),
-                    ],
-                  ),
+                    )
+                  ],),
                 ),
               ],
             ),
@@ -205,13 +178,13 @@ class _RecommendBookCardState extends State<RecommendBookCard> {
         padding: EdgeInsets.only(top: padding),
         child: InkWell(
           onTap: () {
-            goToDetailPage(context, widget.recData.data);
+            goToDetailPageOfBook(context, widget.book);
           },
           child: ContainerWithBoxDecoration(
             child: Padding(
               padding: const EdgeInsets.all(10),
               child: Image.network(
-                widget.recData.data.imgUrl,
+                widget.book.imgUrl,
                 fit: BoxFit.cover,
                 height: imgHeight,
                 width: imgWidth,
@@ -221,7 +194,7 @@ class _RecommendBookCardState extends State<RecommendBookCard> {
         ));
   }
 
-  void goToDetailPage(BuildContext context, Book book) {
+  void goToDetailPageOfBook(BuildContext context, Book book) {
     context.read<BookCubit>().setBook(book);
     context.read<BookCubit>().goToDetailPage(context);
   }
@@ -230,7 +203,7 @@ class _RecommendBookCardState extends State<RecommendBookCard> {
     if (desc.trim().isEmpty) {
       return "- - -";
     }
-    int index = 60;
+    int index = 50;
     String shortDesc = desc;
     if (desc.trim().length > index) {
       shortDesc = shortDesc.replaceAll("\n", " ");
@@ -242,28 +215,21 @@ class _RecommendBookCardState extends State<RecommendBookCard> {
   Text getShortTitle(String title) {
     const int maxChar = 40;
     const int firstLineMaxChar = 20;
-    double fontSize = 18;
+    double fontSize = 20;
     if (title.trim().length > maxChar) {
       title = "${title.substring(0, maxChar).trim()}...";
     }
     if (title.length > firstLineMaxChar) {
-      fontSize = 17;
+      fontSize = 18.5;
     }
 
     Text text = Text(
       title,
       maxLines: 2,
-      style: TextStyle(fontSize: fontSize, fontWeight: FontWeight.bold),
+      style:  TextStyle(fontSize: fontSize, fontWeight: FontWeight.bold),
     );
 
     return text;
   }
-/*
-  String getShortTitle(String title) {
-    int index = 40;
-    if (title.trim().length > index) {
-      return "${title.substring(0, index).trim()}...";
-    }
-    return title;
-  }*/
 }
+*/
