+++ FlutterBookApp/lib/todo.txt
@@ -1,0 +1,7 @@
+login screen duzeltilecek
+
+Recommended BOok Paghe yerine --> book title eklencek
+
+shadowlar azaltilacak
+
+font eklenebilir
